Move LocalAudioLevelIndicator files and remove component from MenuBar
