+++ src/components/MenuBar/DeviceSelector/AudioInputList/AudioInputList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { FormControl, MenuItem, Typography, Select } from '@material-ui/core';
-import LocalAudioLevelIndicator from '../../LocalAudioLevelIndicator/LocalAudioLevelIndicator';
+import LocalAudioLevelIndicator from '../LocalAudioLevelIndicator/LocalAudioLevelIndicator';
 import { makeStyles } from '@material-ui/core/styles';
 import { useAudioInputDevices } from '../hooks/hooks';
 import useVideoContext from '../../../../hooks/useVideoContext/useVideoContext';
