+++ src/components/MenuBar/DeviceSelector/LocalAudioLevelIndicator/LocalAudioLevelIndicator.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { LocalAudioTrack } from 'twilio-video';
-import useVideoContext from '../../../hooks/useVideoContext/useVideoContext';
-import AudioLevelIndicator from '../../AudioLevelIndicator/AudioLevelIndicator';
+import useVideoContext from '../../../../hooks/useVideoContext/useVideoContext';
+import AudioLevelIndicator from '../../../AudioLevelIndicator/AudioLevelIndicator';
 
 export default function LocalAudioLevelIndicator() {
   const { localTracks } = useVideoContext();
